refactor: rename transactions
